get rid of warnings

+++ oglwidget.cpp
@@ -85,7 +85,7 @@ void OGLWidget::resizeGL(int w, int h)
     glViewport(0,0,w,h);
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
-    gluPerspective(45, static_cast<float>(w/h), 0.01, 100.0);
+    gluPerspective(45, static_cast<double>(w/h), 0.01, 100.0);
     glMatrixMode(GL_MODELVIEW);
     glLoadIdentity();
     gluLookAt(0,5,5,
